+++ test/mixedstruct.test.ts
@@ -65,6 +65,30 @@ describe('MixedStruct  test', () => {
     });
 
 
+    it('should success when call unlock', () => {
+
+      let person = new Person({
+        name: new Bytes('7361746f736869206e616b616d6f746f'),
+        addr: new Bytes('68656c6c6f20776f726c6421'),
+        isMale: false,
+        age: 11,
+        blk: new Block({
+          time: 10000,
+          hash: new Bytes('68656c6c6f20776f726c6420'),
+          header: new Bytes('1156'),
+        })
+      });
+
+
+      person.isMale = true;
+      person.age = 33;
+
+      result = mixedStruct.unlock(person).verify()
+      expect(result.success, result.error).to.be.true
+    });
+
+
+
     it('should fail when name error', () => {
       result = mixedStruct.unlock(new Person({
         name: new Bytes('11'),
@@ -158,7 +182,7 @@ describe('MixedStruct  test', () => {
           })]
         }));
 
-      }).to.throw('checkArray fail, struct Bsver property tokens should be struct Token {}[3]');
+      }).to.throw('Member tokens of struct Bsver is of wrong type, expected Token[3]');
 
 
       expect(() => {
@@ -190,7 +214,7 @@ describe('MixedStruct  test', () => {
           })]
         }));
 
-      }).to.throw('checkArray fail, struct Bsver property tokens should be struct Token {}[3]');
+      }).to.throw('Member tokens of struct Bsver is of wrong type, expected Token[3]');
 
 
       expect(() => {
@@ -212,7 +236,7 @@ describe('MixedStruct  test', () => {
             createTime: 1000000
           })
         }));
-      }).to.throw('struct Bsver property tokens should be struct Token {}[3]');
+      }).to.throw('Member tokens of struct Bsver is of wrong type, expected Token[3]');
 
     });
 
@@ -240,6 +264,65 @@ describe('MixedStruct  test', () => {
       result = mixed.unlock(4).verify()
       expect(result.success, result.error).to.be.false
     });
+  })
+})
+
+
+describe('checkstruct read and write field', () => {
+  let bsver;
+
+  before(() => {
+    bsver = new Bsver({
+      name: new Bytes('6666'),
+      friend: new Person({
+        name: new Bytes('7361746f736869206e616b616d6f746f'),
+        addr: new Bytes('6666'),
+        isMale: true,
+        age: 33,
+        blk: new Block({
+          time: 10000,
+          hash: new Bytes('68656c6c6f20776f726c6421'),
+          header: new Bytes('1156'),
+        })
+      }),
+      tokens: [new Token({
+        id: new Bytes('0001'),
+        createTime: 1000000
+      }), new Token({
+        id: new Bytes('0002'),
+        createTime: 1000001
+      }), new Token({
+        id: new Bytes('0003'),
+        createTime: 1000002
+      })]
+    });
 
   })
+
+
+  it('access struct field', () => {
+
+    expect(bsver.name.toLiteral()).to.equal("b'6666'")
+
+    bsver.name = new Bytes("01010101")
+    expect(bsver.name.toLiteral()).to.equal("b'01010101'")
+
+    expect(bsver.friend.name.toLiteral()).to.equal("b'7361746f736869206e616b616d6f746f'")
+
+    bsver.friend.name = new Bytes("1111")
+
+    expect(bsver.friend.name.toLiteral()).to.equal("b'1111'")
+
+    expect(bsver.tokens[0].createTime.toLiteral()).to.equal("1000000")
+
+    bsver.tokens[0].createTime = 33333
+
+    expect(bsver.tokens[0].createTime.toLiteral()).to.equal("33333")
+  });
+
+
+  it('should throw when write struct field with wrong type', () => {
+    expect(() => { bsver.name = 11 }).to.throw('Member name of struct Bsver is of wrong type, expected bytes but got int');
+  })
+
 });