+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+
+## 0.7.0
+
+- support access structure members
+- support the use of commands to compile contracts,  see [command in CLI](./README.md#contract-description-file)
+- change `getStateScript` to `getNewStateScript`
+
+
+*Release Date: 2021/10/23*
+
+
 ## 0.6.2
 
 - support array/struct state